prediction/pyproject.toml
[tool.poetry]
name = "prediction"
version = "0.1.0"
description = ""
authors = ["Alessandro Pacielli <alessandropacielli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
jupyter = "^1.0.0"
numpy = "^1.18.4"
pandas = "^1.0.3"
faust = "^1.10.4"
keras = "^2.3.1"
tensorflow = "^2.2.0"
python-dotenv = "^0.13.0"
joblib = "^0.15.1"
scikit-learn = "^0.23.1"

[tool.poetry.dev-dependencies]
kafka-python = "<1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
